created separate component for ship/bill form-
