+++ client/features/checkout/Checkout.js
@@ -1,36 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
+import ShippingBillingForm from './ShippingBillingForm';
 
 const Checkout = () => {
-
-    const stateAbbreviations = ["AK", "AL", "AR", "AS", "AZ", "CA", "CO", "CT", "DC", "DE", "FL", "GA", "GU", "HI", "IA", "ID", "IL", "IN", "KS", "KY", "LA", "MA", "MD", "ME", "MI", "MN", "MO", "MP", "MS", "MT", "NC", "ND", "NE", "NH", "NJ", "NM", "NV", "NY", "OH", "OK", "OR", "PA", "PR", "RI", "SC", "SD", "TN", "TX", "UM", "UT", "VA", "VI", "VT", "WA", "WI", "WV", "WY"];
-    const stateList = stateAbbreviations.map((state) => {
-        return (
-            <option value={state} key={state}>{state}</option>
-        )
-    })
-
-    const saveShippingInfo = (evt) => {
-        evt.preventDefault();
-        console.log('Shipping details saved!');
-        // Dispatch shipping info to user profile.
-        // Clear out shipping info.
-        setFullName = '';
-        setPhoneNum = '';
-        setAddress = '';
-        setCity = '';
-        setState = '';
-        setZip = '';
-        const cardNameElement = evt.target.children[13];
-        cardNameElement.required = true;
-        setFormTitle('Billing Information')
-        setBillingVisible(false);
-        // POST request to User to save shipping info
-    }
-
-    const saveOrder = (evt) => {
-
-    }
 
     // By submitting the order, we need to...
     // 1. Set 'purchased' on the order to TRUE.
@@ -38,12 +10,7 @@ const Checkout = () => {
     // 3. Reset the user's cart.
     // 4. Adjust each purchased item's stock in the db.
 
-    // Billing form is the same, but with payment method inputs
-
-    const changeUserState = (e) => {
-        setState(e.target.value);
-    }
-
+    
     const sameAsBilling = (e) => { // should this be on the billing page to bring over shipping info?
         if (e.target.checked) {
             // Send shipping info to billing
@@ -51,78 +18,10 @@ const Checkout = () => {
         // If not, clear billing info.
     }
 
-    const [fullName, setFullName] = useState('');
-    const [phoneNum, setPhoneNum] = useState('');
-    const [address, setAddress] = useState('');
-    const [city, setCity] = useState('');
-    const [state, setState] = useState('');
-    const [zip, setZip] = useState('');
-    const [formTitle, setFormTitle] = useState('Shipping Information');
-    const [cardName, setCardName] = useState('');
-    const [billingVisible, setBillingVisible] = useState(true);
-    const [enableSubmit, setEnableSubmit] = useState(true);
-    const [enableSave, setEnableSave] = useState(false);
-
     return (
-        <div>
-            <h1>Checkout</h1>
-        <form id="shipping-billing-info" onSubmit={saveShippingInfo}>
-            <h2>{formTitle}</h2>
-            <label htmlFor='fullName'> Full name: </label>
-            <input
-                name="fullName"
-                value={fullName}
-                required={true}
-                placeholder='First and Last name'
-                onChange={(e) => setFullName(e.target.value)}
-                />
-            <label htmlFor='phoneNum'> Phone Number: </label>
-            <input
-                name="phoneNum"
-                value={phoneNum}
-                required={true}
-                onChange={(e) => setPhoneNum(e.target.value)}
-                />
-            <label htmlFor='address'> Address: </label>
-            <input
-                name="address"
-                value={address}
-                required={true}
-                placeholder='Street address or P.O. Box'
-                onChange={(e) => setAddress(e.target.value)}
-                />
-            <label htmlFor='city'> City: </label>
-            <input
-                name="city"
-                value={city}
-                required={true}
-                onChange={(e) => setCity(e.target.value)}
-                />
-            <select value={state} onChange={changeUserState}>
-                {stateList}
-            </select>
-            <label htmlFor='zip'> Zip Code: </label>
-            <input
-                name="zip"
-                value={zip}
-                required={true}
-                // Valid zipcode/numbers only 
-                onChange={(e) => setZip(e.target.value)}
-                />
-            <label htmlFor='cardName'>Name on Card:</label>
-            <input
-                name='cardName'
-                value={cardName}
-                // required={true}
-                hidden={billingVisible}
-                onChange={(e) => setCardName(e.target.value)}
-                />
-                {/* <label htmlFor="billingInfo"> Same as billing? </label> */}
-                {/* <input type="checkbox" id="billingInfo" value="billingInfoSame" onClick={sameAsBilling}/> */}
-                <button type='submit'>Submit Order</button>
-        </form>
-        </div>
+        <ShippingBillingForm />
     )
+    
 }
 
 export default Checkout;
